EES-2572 - refactor misc test utils to use common services
