+++ useful-scripts/misc-test-utils/src/index.ts
@@ -1,9 +1,9 @@
 import inquirer from 'inquirer';
 import chalk from 'chalk';
 import 'dotenv-safe/config';
-import createReleaseAndPublish from './modules/publication/publishPublication';
-import createPublicationAndRelease from './modules/release/createRelease';
 import uploadSingleSubject from './modules/subject/uploadSubject';
+import createPublicationAndRelease from './modules/release/createRelease';
+import createReleaseAndPublish from './modules/publication/publishPublication';
 
 process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0';
