+++ useful-scripts/misc-test-utils/src/utils/zipDirectory.ts
@@ -1,5 +1,3 @@
-/* eslint-disable */
-
 import archiver from 'archiver';
 import fs from 'fs';
